requirements.txt
Flask==1.0.1
pyserial==3.4
